Fix encoding issue for query path value
In buildUrl, encoding of the values for each
of the query parameters was done when appending them
to the searchParams. However, since buildUrl also
decoded such encoding, the json query path had special
characters decoded.

To have the json query path encoded, we encode it once
prior to calling the buildUrl function.